requirements.txt
# Web scraping and browser automation
playwright

# HTML parsing
beautifulsoup4

# Azure Cognitive Search
azure-core
azure-search-documents

# Neo4j connection
neo4j

# Environment variables
python-dotenv

# OpenAI embeddings
openai

# Token counting
tiktoken

# HTTP requests
requests